+++ ProyectoJavaCERO/CERO/Negocio/src/main/java/com/mycompany/negocio/dtos/NuevoPagoDTO.java
@@ -1,25 +1,27 @@
 
 package com.mycompany.negocio.dtos;
 
+import java.math.BigDecimal;
+
 /**
  * Clase de transporte que representa un nuevo pago de una inscripción.
  * @author victoria
  */
 public class NuevoPagoDTO {
     
-    private Float total;
+    private BigDecimal total;
     private MetodoPagoDTO metodoPago;
 
-    public NuevoPagoDTO(Float total, MetodoPagoDTO metodoPago) {
+    public NuevoPagoDTO(BigDecimal total, MetodoPagoDTO metodoPago) {
         this.total = total;
         this.metodoPago = metodoPago;
     }
 
-    public Float getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
 
-    public void setTotal(Float total) {
+    public void setTotal(BigDecimal total) {
         this.total = total;
     }
 
@@ -30,6 +32,8 @@ public class NuevoPagoDTO {
     public void setMetodoPago(MetodoPagoDTO metodoPago) {
         this.metodoPago = metodoPago;
     }
+
+   
     
     
 }
